+++ hw1/simpleCharManager.cpp
@@ -16,7 +16,8 @@ simpleCharManager::~simpleCharManager(){
 }
              
 char* simpleCharManager::alloc_chars(int n){
-	int max_alloc_len = ((int)( &(buffer[BUF_SIZE-1]) - free_place) + sizeof(char))/sizeof(char);
+	int max_alloc_len = ((int)( &(buffer[BUF_SIZE-1]) - 
+		free_place) + (int)sizeof(char))/(int)sizeof(char);
 
 	// alloc if space is availible in buffer
 	if (n > 0 && n <= max_alloc_len) {
@@ -24,13 +25,13 @@ char* simpleCharManager::alloc_chars(int n){
 		free_place = free_place + n;
 		return alloc_ptr;
 	} else {
-		// std::cerr << "Error: invalid memory allocation index" << std::endl;
         return NULL;
 	}
 }
 
 void simpleCharManager::free_chars(char* p){
 
+	// check if p is within allocated memory
 	if (p<free_place && p >= buffer) {
 		char* old_free_place = free_place;
 		free_place = p;
@@ -43,6 +44,5 @@ void simpleCharManager::free_chars(char* p){
 	}
 
 	else {
-		// std::cerr << "Error: cannot free memory that has not been allocated" << std::endl;
 	}
 }
